Bibliographic data base creation

db_bib.wos.data() converts a data.frame with WoS variables (as returned by import_wos; see wosdf) into a relational database (a list of data.frames).

Usage

db_bib(data, ...)

# S3 method for class 'wos.data'
db_bib(data, label = "", progress = interactive(), verbose = FALSE, ...)

# S3 method for class 'wos.db'
print(x, ...)

# S3 method for class 'wos.db'
plot(
  x,
  filter,
  which = 1:3,
  plot = TRUE,
  warning = FALSE,
  ask = plot && (length(which) > 1) && interactive(),
  ...
)

Arguments

data

a data.frame with bibliographic data.

...

further arguments passed to or from other methods.

label

character string describing the data.

progress

logical; if TRUE the progress is printed.

verbose

logical; if TRUE additional information is printed.

x

an object used to select a method.

filter

vector of document identifiers (usually a result of get_id_docs).

which

if a subset of the plots is required, specify a subset of the numbers 1:3.

plot

logical; if TRUE (default), the plots are drawn, otherwise only the list of ggplot2 objects is (invisibly) returned.

warning

logical; if FALSE (default), warnings are ignored.

ask

logical; if TRUE, the user is asked before each plot (see par(ask=.)).

Value

An S3 object of class wos.db. A list with the following components:

Docs: document-level records (year, type, citations, identifiers).

Authors and AutDoc: author dictionary and document-author links.

OI and OIDoc: ORCID identifiers and their linkage to documents.

RI and RIDoc: ResearcherID identifiers and their linkage to documents.

Affiliations and AffDoc: affiliation dictionary and document links.

Addresses and AddAutDoc: address and document-author-address links.

Sources: document sources (journals, books, proceedings, ...).

Categories, Areas, CatSour and AreaSour: thematic classifications and their linkage to sources.

WSIndex and SourWSI: Web of Science indexes and their linkage to sources.

label and date: metadata that helps tracking the dataset identity over time.

This object has specialized print and plot methods.

See also

Examples

db <- db_bib(wosdf)
str(db, 1)
#> List of 20
#>  $ Docs        :'data.frame':	293 obs. of  28 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:28] "Document ID" "Source ID" "Article Title" "Document Type" ...
#>   .. ..- attr(*, "names")= chr [1:28] "idd" "ids" "TI" "DT" ...
#>  $ Authors     :'data.frame':	504 obs. of  5 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:5] "Author ID" "Author" "Author Full Name" "ORCID ID" ...
#>   .. ..- attr(*, "names")= chr [1:5] "ida" "AU" "AF" "ioi" ...
#>  $ AutDoc      :'data.frame':	1065 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "Author ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "ida"
#>  $ OI          :'data.frame':	262 obs. of  3 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:3] "ORCID ID" "ORCID" "ORCID Full Name"
#>   .. ..- attr(*, "names")= chr [1:3] "ioi" "OI" "AFOI"
#>  $ OIDoc       :'data.frame':	695 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "ORCID ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "ioi"
#>  $ RI          :'data.frame':	247 obs. of  3 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:3] "RI ID" "Researcher Ids" "RI Full Name"
#>   .. ..- attr(*, "names")= chr [1:3] "iri" "RI" "AFRI"
#>  $ RIDoc       :'data.frame':	734 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "RI ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "iri"
#>  $ Categories  :'data.frame':	254 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "id WOS category" "WOS category"
#>   .. ..- attr(*, "names")= chr [1:2] "idc" "WC"
#>  $ CatSour     :'data.frame':	228 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Source ID" "Categorie ID"
#>   .. ..- attr(*, "names")= chr [1:2] "ids" "idc"
#>  $ Areas       :'data.frame':	26 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Area ID" "Research Area"
#>   .. ..- attr(*, "names")= chr [1:2] "idra" "SC"
#>  $ AreaSour    :'data.frame':	198 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Source ID" "Area ID"
#>   .. ..- attr(*, "names")= chr [1:2] "ids" "idra"
#>  $ Addresses   :'data.frame':	843 obs. of  5 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:5] "Address ID" "Document ID" "Address" "University" ...
#>   .. ..- attr(*, "names")= chr [1:5] "idad" "idd" "C1" "Univ" ...
#>  $ AddAutDoc   :'data.frame':	1275 obs. of  3 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:3] "Author ID" "Document ID" "Address ID"
#>   .. ..- attr(*, "names")= chr [1:3] "ida" "idd" "idad"
#>  $ Affiliations:'data.frame':	148 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Affiliation ID" "Affiliation"
#>   .. ..- attr(*, "names")= chr [1:2] "idaf" "C3"
#>  $ AffDoc      :'data.frame':	673 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "Affiliation ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "idaf"
#>  $ Sources     :'data.frame':	118 obs. of  14 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:17] "Source ID" "WoS Category" "Research Area" "Web of Science Index" ...
#>   .. ..- attr(*, "names")= chr [1:17] "ids" "WC" "SC" "WE" ...
#>  $ WSIndex     :'data.frame':	3 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "WoS Index ID" "Web of Science Index"
#>   .. ..- attr(*, "names")= chr [1:2] "idwe" "WE"
#>  $ SourWSI     :'data.frame':	133 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Source ID" "WoS Index ID"
#>   .. ..- attr(*, "names")= chr [1:2] "ids" "idwe"
#>  $ label       : chr ""
#>  $ date        : chr "2025-02-19"
#>  - attr(*, "class")= chr "wos.db"
print(db)
#> List of 20
#>  $ Docs        :'data.frame':	293 obs. of  28 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:28] "Document ID" "Source ID" "..
#>   .. ..- attr(*, "names")= chr [1:28] "idd" "ids" "TI" "DT" ...
#>  $ Authors     :'data.frame':	504 obs. of  5 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:5] "Author ID" "Author" "Autho"..
#>   .. ..- attr(*, "names")= chr [1:5] "ida" "AU" "AF" "ioi" ...
#>  $ AutDoc      :'data.frame':	1065 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "Author ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "ida"
#>  $ OI          :'data.frame':	262 obs. of  3 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:3] "ORCID ID" "ORCID" "ORCID F"..
#>   .. ..- attr(*, "names")= chr [1:3] "ioi" "OI" "AFOI"
#>  $ OIDoc       :'data.frame':	695 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "ORCID ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "ioi"
#>  $ RI          :'data.frame':	247 obs. of  3 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:3] "RI ID" "Researcher Ids" "R"..
#>   .. ..- attr(*, "names")= chr [1:3] "iri" "RI" "AFRI"
#>  $ RIDoc       :'data.frame':	734 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "RI ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "iri"
#>  $ Categories  :'data.frame':	254 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "id WOS category" "WOS cate"..
#>   .. ..- attr(*, "names")= chr [1:2] "idc" "WC"
#>  $ CatSour     :'data.frame':	228 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Source ID" "Categorie ID"
#>   .. ..- attr(*, "names")= chr [1:2] "ids" "idc"
#>  $ Areas       :'data.frame':	26 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Area ID" "Research Area"
#>   .. ..- attr(*, "names")= chr [1:2] "idra" "SC"
#>  $ AreaSour    :'data.frame':	198 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Source ID" "Area ID"
#>   .. ..- attr(*, "names")= chr [1:2] "ids" "idra"
#>  $ Addresses   :'data.frame':	843 obs. of  5 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:5] "Address ID" "Document ID" "..
#>   .. ..- attr(*, "names")= chr [1:5] "idad" "idd" "C1" "Univ" ...
#>  $ AddAutDoc   :'data.frame':	1275 obs. of  3 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:3] "Author ID" "Document ID" ""..
#>   .. ..- attr(*, "names")= chr [1:3] "ida" "idd" "idad"
#>  $ Affiliations:'data.frame':	148 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Affiliation ID" "Affiliation"
#>   .. ..- attr(*, "names")= chr [1:2] "idaf" "C3"
#>  $ AffDoc      :'data.frame':	673 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Document ID" "Affiliation ID"
#>   .. ..- attr(*, "names")= chr [1:2] "idd" "idaf"
#>  $ Sources     :'data.frame':	118 obs. of  14 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:17] "Source ID" "WoS Category""..
#>   .. ..- attr(*, "names")= chr [1:17] "ids" "WC" "SC" "WE" ...
#>  $ WSIndex     :'data.frame':	3 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "WoS Index ID" "Web of Scie"..
#>   .. ..- attr(*, "names")= chr [1:2] "idwe" "WE"
#>  $ SourWSI     :'data.frame':	133 obs. of  2 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:2] "Source ID" "WoS Index ID"
#>   .. ..- attr(*, "names")= chr [1:2] "ids" "idwe"
#>  $ label       : chr ""
#>  $ date        : chr "2025-02-19"
#>  - attr(*, "class")= chr "wos.db"
summary(db)
#> Number of documents: 293 
#> Authors: 504 
#> Period: 2018 - 2023 
#> 
#> Document types:
#>                            Documents
#> Article                          254
#> Article; Early Access              1
#> Article; Proceedings Paper         2
#> Correction                         7
#> Editorial Material                21
#> Review                             8
#> 
#> Number of authors per document:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   3.000   3.000   3.635   4.000  13.000 
#> 
#> Number of documents per author:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   1.000   1.000   2.113   2.000  23.000 
#> 
#> Number of times cited:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.00    1.00    3.00    5.87    7.00  123.00 
#> 
#> Indexes:
#>  H  G 
#> 19 27 
#> 
#> Status:
#> Highly Cited   Hot Papers 
#>            1            0 
#> 
#> Top Categories:
#>                                                  Documents
#> Mathematics, Applied                                   133
#> Mathematics                                             96
#> Statistics & Probability                                84
#> Mathematics, Interdisciplinary Applications             55
#> Computer Science, Interdisciplinary Applications        30
#> Logic                                                   28
#> Mechanics                                               19
#> Engineering, Multidisciplinary                          17
#> Multidisciplinary Sciences                              13
#> Biochemical Research Methods                            11
#> Others                                                 104
#> 
#> Top Areas:
#>                                          Documents
#> Mathematics                                    293
#> Computer Science                                45
#> Science & Technology - Other Topics             41
#> Engineering                                     20
#> Mechanics                                       19
#> Biochemistry & Molecular Biology                11
#> Mathematical & Computational Biology             9
#> Operations Research & Management Science         9
#> Physics                                          9
#> Mathematical Methods In Social Sciences          7
#> Others                                          37
#> 
#> Top Journals:
#>                                          Documents
#> Log. J. IGPL                                    28
#> Mathematics                                     21
#> Test                                            17
#> Appl. Math. Comput.                              9
#> Complexity                                       9
#> IEEE-ACM Trans. Comput. Biol. Bioinform.         7
#> J. Math. Anal. Appl.                             7
#> Comput. Math. Appl.                              6
#> J. Comput. Appl. Math.                           6
#> Arch. Comput. Method Eng.                        5
#> Others                                         178
#> 
#> Top Countries:
#>           Documents
#> Spain           293
#> France           20
#> UK               18
#> USA              17
#> Italy            13
#> Chile             9
#> Ecuador           9
#> Portugal          8
#> Argentina         6
#> Uruguay           6
#> Others           46
#> 
#> International colaboration documents (Spain): 134