Import JCR data from Web of Science (WoS)

Reads JCR data from excel files downloaded from WoS and generates a relational database (a list of data.frames). It is assumed that the file name format is JCR_[WE]_[PY], where [WE] is the WoS index (SCIE, SSCI, ...) and [PY] the JCR year.

Usage

db_jcr(path = ".", files = dir(path, pattern = "*.xlsx"), verbose = TRUE)

Arguments

path

character; path to the directory containing the files. Defaults to the working directory.

files

character vector with the file names. Defaults to filenames in directory path with extension .xlsx".

verbose

logical; indicating whether the name of the file being processed is printed. Defaults to TRUE.

Value

An S3 object of class jcr.db. A list with components: Sources, Categories, JCRSour and JCRCatSour.

Details

For successful execution, Excel files must retain the original header fields from Clarivate exports: Title20, ISO_ABBREV, TITLE, ISSN, EISSN, ISSUES/YEAR, COUNTRY, LANGUAGE, 1ST_YR_PUB, categories, TOT_CITES, CITES_JCR_YR, CITES_JCR_YR1, CITES_JCR_YR2, IF_NUMERATOR, CITES_JCR_YR3, CITES_JCR_YR4, CITES_JCR_YR5, 5YR_IF_NUMERATOR, ITEMS_JCR_YR, ITEMS_JCR_YR1, ITEMS_JCR_YR2, IF_DENOMINATOR, ITEMS_JCR_YR3, ITEMS_JCR_YR4, ITEMS_JCR_YR5, 5YR_IF_DENOMINATOR, IMPACT_FACTOR, IMMEDIACY_INDEX, CITED_HALF_LIFE, 5YR_IMPACT_FACTOR, EIGENFACTOR, NORM_EIGENFACTOR, ARTL_INFLUENCE, FREQUENCY, PUBCODE, PUBLISHER_NAME, CATEGORY_CODE, CATEGORY_DESCRIPTION, CATEGORY_RANKING, QUARTILE_RANK, JIF_PERCENTILE.

See also