Summaries of a bibliographic database

summary methods for WoS bibliometric databases.

Usage

# S3 method for class 'wos.db'
summary(object, filter, TC.index = c("H", "G"), nmax = 10, ...)

# S3 method for class 'summary.wos.db'
print(x, ...)

TC.authors(db, idAuthors)

# S3 method for class 'summary.wos.db'
plot(
  x,
  which = 1:5,
  pie = FALSE,
  others = !pie,
  accuracy = 0.1,
  plot = TRUE,
  ask = length(which) > 1 && interactive(),
  ...
)

# S3 method for class 'wos.jcr'
summary(object, filter, all = FALSE, ...)

# S3 method for class 'summary.wos.jcr'
print(x, digits = 2, all = TRUE, ...)

# S3 method for class 'summary.wos.jcr'
plot(x, plot = TRUE, all = FALSE, ask = plot && all && interactive(), ...)

Arguments

object

an object for which a summary is desired.

filter

vector of document identifiers (optional, usually a result of get.idDocs).

TC.index

citation indexes.

nmax

number of top levels.

...

further arguments passed to or from other methods.

x

an object used to select a method.

db

Object of class wos.db, as returned by CreateDB.

idAuthors

optional; author identifiers

which

if a subset of the plots is required, specify a subset of the numbers 1:5.

pie

logical; if TRUE, pie charts are drawn.

others

logical; if FALSE, only nmax top levels categories are shown.

accuracy

numerical scalar; see label_number(accuracy=.).

plot

logical; if TRUE (default), the plots are drawn, otherwise only the list of ggplot2 objects is (invisibly) returned.

ask

logical; if TRUE, the user is asked before each plot, see par(ask=.).

all

logical; if TRUE, the corresponding wos.db method is called (additional parameters ... are passed to it), otherwise only results of JCR metrics are returned.

digits

minimal number of significant digits, see print.default.

See also