Summaries per year

The summary_year methods return summaries by year of a bibliometric database.

Usage

summary_year(object, ...)

# S3 method for class 'wos.db'
summary_year(object, filter, ...)

# S3 method for class 'summary.year.wos'
print(x, ...)

# S3 method for class 'summary.year.wos'
plot(
  x,
  which = 1:3,
  boxplot = FALSE,
  plot = TRUE,
  ask = plot && length(which) > 1 && interactive(),
  ...
)

# S3 method for class 'wos.jcr'
summary_year(object, filter, all = FALSE, ...)

# S3 method for class 'summary.year.jcr'
print(x, digits = 3, all = TRUE, ...)

# S3 method for class 'summary.year.jcr'
plot(x, plot = TRUE, all = FALSE, ask = plot && interactive(), ...)

Arguments

object

an object for which a summary is desired.

...

further arguments passed to or from other methods.

filter

vector of document identifiers (optional, usually a result of get_id_docs).

x

an object used to select a method.

which

if a subset of the plots is required, specify a subset of the numbers 1:3.

boxplot

logical; if TRUE, boxplots are drawn (for plots from 2 to 3).

plot

logical; if TRUE (default), the plots are drawn, otherwise only the list of ggplot2 objects is (invisibly) returned.

ask

logical; if TRUE, the user is asked before each plot, see par(ask=.).

all

logical; if TRUE, the corresponding wos.db method is called (additional parameters ... are passed to it), otherwise only results of JCR metrics are returned.

digits

minimal number of significant digits, see print.default.

Value

A list of summary statistics (an object of class summary.year.wos and/or summary.year.jcr), which has specialized print() and plot() methods.

See also

Examples

# Bibliographic database
db <- db_bib(wosdf)
summary_year(db)
#> Annual Scientific Production:
#>      Documents
#> 2018        40
#> 2019        61
#> 2020        46
#> 2021        51
#> 2022        45
#> 2023        50
#> 
#> Annual Authors per Document:
#>    PY     Mean Median
#>  2018 3.325000    3.5
#>  2019 3.590164    3.0
#>  2020 3.869565    3.0
#>  2021 3.509804    3.0
#>  2022 3.844444    3.0
#>  2023 3.660000    3.0
#> 
#> Annual Times Cited:
#>    PY Cites     Mean Median
#>  2018   364 9.100000    5.0
#>  2019   538 8.819672    5.0
#>  2020   306 6.652174    3.0
#>  2021   206 4.039216    3.0
#>  2022   186 4.133333    2.0
#>  2023   120 2.400000    1.5
#> 
#> Status:
#>    PY Highly Cited Hot Paper
#>  2018            0         0
#>  2019            1         0
#>  2020            0         0
#>  2021            0         0
#>  2022            0         0
#>  2023            0         0
# Bibliographic database with JCR metrics
summary_year(dbjcr)
#> Mean of JCR metrics:
#>    PY  JIF   IMM  CHL JIF5     JEF JEFN   JAI JIFP   WCP
#>  2018 2.01 0.579 8.03 2.11 0.01397 1.66 0.864 54.8  89.7
#>  2019 1.60 0.527 8.85 1.61 0.00876 1.07 0.605 49.3  78.4
#>  2020 2.59 0.841 7.09 2.68 0.01030 2.16 0.899 60.8  87.2
#>  2021 2.27 0.624 8.81 2.92 0.01108 2.38 1.068 53.7  73.5
#>  2022 2.28 0.638 8.98 2.39 0.01050 2.29 0.769 55.3  85.2
#>  2023 1.98 0.520 9.38 2.21 0.01349 2.97 0.830 58.4 103.7
#> 
#> Documents in JCR:
#> 2018 2019 2020 2021 2022 2023 
#>   40   61   45   51   45   50 
#> 
#> Documents in Q1:
#> 2018 2019 2020 2021 2022 2023 
#>   13   13   19   15   13   17 
#> 
#> Documents in D1:
#> 2018 2019 2020 2021 2022 2023 
#>    6    5   15    8    7    9 
#> 
#> Documents in top 3 journals:
#> 2018 2019 2020 2021 2022 2023 
#>    0    0    0    0    0    0