WoS bibliographic data base

wos_CreateDB converts a data.frame with WoS variables (see wosdf) into...

wos_CreateDB(wosdf, label = "", print = interactive())

# S3 method for wos.db
print(x, ...)

# S3 method for wos.db
plot(x, filter, which = 1:3, plot.summary = FALSE,
  which.summary = 1:5, nmax = 10, plot.summary.year = FALSE,
  which.summary.year = 1:3, ask = (length(which) + plot.summary *
  length(which.summary) + plot.summary.year * length(which.summary.year) >
  1) && dev.interactive(), ...)

Arguments

wosdf
data.frame with WoS data... (as returned by wos_ImportSources).

print
logical; if TRUE the progress is printed.

x
an object used to select a method.

...
further arguments passed to or from other methods.

filter
vector of document identifiers (usually a result of get.idDocs).

which
if a subset of the plots is required, specify a subset of the numbers 1:3.

plot.summary
logical; if TRUE, plot.summary.wos.db is called.

which.summary
indices of the required plots in plot.summary.wos.db.

nmax
number of top levels (see summary.wos.db).

plot.summary.year
logical; if TRUE, plot.summary.year is called.

which.summary.year
indices of the required plots in plot.summary.year.

ask
logical; if TRUE, the user is asked before each plot, see par(ask=.).

Value

An S3 object of class wos.db. A list with the folowing data.frames:

Docs

Authors AutDoc Categories CatDoc Areas AreaDoc Addresses AddAutDoc Journals

See also

Examples

db <- wos_CreateDB(wosdf)
str(db, 1)

#> List of 11
#>  $ Docs      :'data.frame':	856 obs. of  26 variables:
#>  $ Authors   :'data.frame':	4051 obs. of  4 variables:
#>  $ AutDoc    :'data.frame':	5511 obs. of  2 variables:
#>  $ Categories:'data.frame':	189 obs. of  2 variables:
#>  $ CatDoc    :'data.frame':	1495 obs. of  2 variables:
#>  $ Areas     :'data.frame':	121 obs. of  2 variables:
#>  $ AreaDoc   :'data.frame':	1364 obs. of  2 variables:
#>  $ Addresses :'data.frame':	3655 obs. of  5 variables:
#>  $ AddAutDoc :'data.frame':	7751 obs. of  3 variables:
#>  $ Journals  :'data.frame':	520 obs. of  12 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:62] "Publication type" "Author" "Book authors" "Editor" ...
#>   .. ..- attr(*, "names")= chr [1:62] "PT" "AU" "BA" "BE" ...
#>  $ label     : chr ""
#>  - attr(*, "class")= chr "wos.db"

print(db)

#> List of 11
#>  $ Docs      :'data.frame':	856 obs. of  26 variables:
#>  $ Authors   :'data.frame':	4051 obs. of  4 variables:
#>  $ AutDoc    :'data.frame':	5511 obs. of  2 variables:
#>  $ Categories:'data.frame':	189 obs. of  2 variables:
#>  $ CatDoc    :'data.frame':	1495 obs. of  2 variables:
#>  $ Areas     :'data.frame':	121 obs. of  2 variables:
#>  $ AreaDoc   :'data.frame':	1364 obs. of  2 variables:
#>  $ Addresses :'data.frame':	3655 obs. of  5 variables:
#>  $ AddAutDoc :'data.frame':	7751 obs. of  3 variables:
#>  $ Journals  :'data.frame':	520 obs. of  12 variables:
#>   ..- attr(*, "variable.labels")= Named chr [1:62] "Publication type" "Author" "Book authors" "Editor" ...
#>   .. ..- attr(*, "names")= chr [1:62] "PT" "AU" "BA" "BE" ...
#>  $ label     : chr ""
#>  - attr(*, "class")= chr "wos.db"

summary(db)

#> Number of documents: 856 
#> Authors: 4051 
#> Period: 2015 
#> 
#> Document types:
#>                    Documents
#> Article                  618
#> Book Review                7
#> Correction                 4
#> Editorial Material        22
#> Letter                    24
#> Meeting Abstract         114
#> News Item                  2
#> Proceedings Paper         28
#> Review                    37
#> 
#> Number of authors per document:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   3.000   5.000   6.438   8.000  98.000 
#> 
#> Number of documents per author:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    1.00    1.00    1.00    1.36    1.00   16.00 
#> 
#> Number of times cited:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.000   0.000   1.000   2.071   3.000  28.000 
#> 
#> Indexes:
#>  H  G 
#> 14 17 
#> 
#> Top Categories:
#>                                           Documents
#> Rheumatology                                     48
#> Cardiac & Cardiovascular Systems                 40
#> Environmental Sciences                           39
#> Engineering, Civil                               35
#> Plant Sciences                                   31
#> Surgery                                          31
#> Oncology                                         26
#> Respiratory System                               26
#> Chemistry, Analytical                            25
#> Computer Science, Artificial Intelligence        25
#> Others                                         1169
#> 
#> Top Areas:
#>                                    Documents
#> Engineering                               87
#> Chemistry                                 84
#> Computer Science                          65
#> Environmental Sciences & Ecology          50
#> Rheumatology                              48
#> Cardiovascular System & Cardiology        43
#> Mathematics                               38
#> Biochemistry & Molecular Biology          32
#> Business & Economics                      32
#> Plant Sciences                            31
#> Others                                   854
#> 
#> Top Journals:
#>                               Documents
#> Arthritis Rheumatol.                 21
#> Osteoarthritis Cartilage             14
#> Rev. Esp. Cardiol.                   10
#> PLoS One                             10
#> Commun. Soil Sci. Plant Anal.         9
#> Inorg. Chem.                          9
#> Eur. J. Phycol.                       9
#> J. Heart Lung Transplant.             9
#> Constr. Build. Mater.                 8
#> Sci. Total Environ.                   6
#> Others                              751
#> 
#> Top Countries:
#>             Documents
#> Spain             855
#> USA                71
#> UK                 62
#> France             51
#> Portugal           49
#> Italy              38
#> Germany            34
#> Belgium            24
#> Netherlands        23
#> Brazil             15
#> Others            236