Summaries of a bibliographic database

summary() methods for bibliometric databases.

Usage

# S3 method for class 'wos.db'
summary(object, filter, index = c("H", "G"), nmax = 10, ...)

# S3 method for class 'summary.wos.db'
print(x, ...)

author_metrics(db, id_authors)

# S3 method for class 'summary.wos.db'
plot(
  x,
  which = 1:5,
  pie = FALSE,
  others = !pie,
  accuracy = 0.1,
  plot = TRUE,
  ask = length(which) > 1 && interactive(),
  ...
)

# S3 method for class 'wos.jcr'
summary(object, filter, all = FALSE, ...)

# S3 method for class 'summary.wos.jcr'
print(x, digits = 2, all = TRUE, ...)

# S3 method for class 'summary.wos.jcr'
plot(x, plot = TRUE, all = FALSE, ask = plot && all && interactive(), ...)

Arguments

object

an object for which a summary is desired.

filter

vector of document identifiers (optional, usually a result of get_id_docs()).

index

citation indexes.

nmax

number of top levels.

...

further arguments passed to or from other methods.

x

an object used to select a method.

db

Object of class wos.db, as returned by db_bib.

id_authors

optional; author identifiers

which

if a subset of the plots is required, specify a subset of the numbers 1:5.

pie

logical; if TRUE, pie charts are drawn.

others

logical; if FALSE, only nmax top levels categories are shown.

accuracy

numerical scalar; see label_number(accuracy=.).

plot

logical; if TRUE (default), the plots are drawn, otherwise only the list of ggplot2 objects is (invisibly) returned.

ask

logical; if TRUE, the user is asked before each plot, see par(ask=.).

all

logical; if TRUE, the corresponding wos.db method is called (additional parameters ... are passed to it), otherwise only results of JCR metrics are returned.

digits

minimal number of significant digits, see print.default.

Value

A list of summary statistics (an object of class summary.wos.db and/or summary.wos.jcr), which has specialized print() and plot() methods.

See also

Examples

# Bibliographic database
db <- db_bib(wosdf)
summary(db)
#> Number of documents: 293 
#> Authors: 504 
#> Period: 2018 - 2023 
#> 
#> Document types:
#>                            Documents
#> Article                          254
#> Article; Early Access              1
#> Article; Proceedings Paper         2
#> Correction                         7
#> Editorial Material                21
#> Review                             8
#> 
#> Number of authors per document:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   3.000   3.000   3.635   4.000  13.000 
#> 
#> Number of documents per author:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   1.000   1.000   2.113   2.000  23.000 
#> 
#> Number of times cited:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.00    1.00    3.00    5.87    7.00  123.00 
#> 
#> Indexes:
#>  H  G 
#> 19 27 
#> 
#> Status:
#> Highly Cited   Hot Papers 
#>            1            0 
#> 
#> Top Categories:
#>                                                  Documents
#> Mathematics, Applied                                   133
#> Mathematics                                             96
#> Statistics & Probability                                84
#> Mathematics, Interdisciplinary Applications             55
#> Computer Science, Interdisciplinary Applications        30
#> Logic                                                   28
#> Mechanics                                               19
#> Engineering, Multidisciplinary                          17
#> Multidisciplinary Sciences                              13
#> Biochemical Research Methods                            11
#> Others                                                 104
#> 
#> Top Areas:
#>                                          Documents
#> Mathematics                                    293
#> Computer Science                                45
#> Science & Technology - Other Topics             41
#> Engineering                                     20
#> Mechanics                                       19
#> Biochemistry & Molecular Biology                11
#> Mathematical & Computational Biology             9
#> Operations Research & Management Science         9
#> Physics                                          9
#> Mathematical Methods In Social Sciences          7
#> Others                                          37
#> 
#> Top Journals:
#>                                          Documents
#> Log. J. IGPL                                    28
#> Mathematics                                     21
#> Test                                            17
#> Appl. Math. Comput.                              9
#> Complexity                                       9
#> IEEE-ACM Trans. Comput. Biol. Bioinform.         7
#> J. Math. Anal. Appl.                             7
#> Comput. Math. Appl.                              6
#> J. Comput. Appl. Math.                           6
#> Arch. Comput. Method Eng.                        5
#> Others                                         178
#> 
#> Top Countries:
#>           Documents
#> Spain           293
#> France           20
#> UK               18
#> USA              17
#> Italy            13
#> Chile             9
#> Ecuador           9
#> Portugal          8
#> Argentina         6
#> Uruguay           6
#> Others           46
#> 
#> International colaboration documents (Spain): 134
# Bibliographic database with JCR metrics
summary(dbjcr)
#> JCR metrics:
#>         JIF  IMM  CHL  JIF5     JEF   JEFN  JAI JIFP WCP
#> Min.    0.4 0.00  1.1  0.48 0.00042  0.053 0.18  2.8   5
#> 1st Qu. 1.0 0.26  5.6  1.00 0.00226  0.381 0.42 28.6  31
#> Median  1.7 0.45  7.7  1.87 0.00500  0.848 0.66 56.3  58
#> Mean    2.1 0.62  8.6  2.30 0.01119  2.060 0.83 55.1  86
#> 3rd Qu. 2.6 0.83 10.4  2.73 0.01412  2.352 0.89 82.6 126
#> Max.    9.7 2.69 39.5 11.81 0.20530 29.627 4.50 97.6 298
#> NA's    1.0 1.00  1.0  2.00 1.00000  1.000 2.00  1.0   1
#> 
#> Documents in JCR: 292 (99.66%)
#> Documents in Q1: 90 (30.82% of JCR)
#> Documents in D1: 50 (17.12% of JCR)
#> Documents in top 3 journals: 0 (0% of JCR)