WoS DB Summaries

summary methods for wos.db-class...

# S3 method for wos.db
summary(object, filter, TC.index = c("H", "G"), nmax = 10, ...)

# S3 method for summary.wos.db
print(x, ...)

TC.authors(db, idAuthors)

# S3 method for summary.wos.db
plot(
  x,
  which = 1:5,
  pie = FALSE,
  others = !pie,
  accuracy = 0.1,
  ask = length(which) > 1 && dev.interactive(),
  ...
)

Arguments

object
an object for which a summary is desired.

filter
vector of document identifiers (usually a result of get.idDocs).

TC.index
citation indexes.

nmax
number of top levels.

...
further arguments passed to or from other methods.

x
an object used to select a method.

db
Object of class wos.db, as returned by CreateDB.wos.

idAuthors
optional; author identifiers

which
if a subset of the plots is required, specify a subset of the numbers 1:5.

pie
logical; if TRUE, pie charts are drawn.

others
logical; if FALSE, only nmax top levels categories are shown.

accuracy
numerical scalar; see label_number(accuracy=.).

ask
logical; if TRUE, the user is asked before each plot, see par(ask=.).

See also